Moved nvml code to separate subdir, modifed wrapnvml to work without cuda. (We want to use nvml for nvidia opencl too)
